refactor🔨: profile tabs
