+++ lib/contexts/profile/presentation/organisms/organisms.dart
@@ -1,0 +1,2 @@
+export "./received_praises_tab.dart";
+export "./communities_tab.dart";
